MangaPill fix
Removed url encoding as the source doesn't seems to need it.
Fixed genres view not working due to the url path being incorrect (added a regex to remove all spaces with a + sign)
Fixed view more not working due to the metadata not being properly accessed.
Incremented version from 2.0.9 to 2.1.0